requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
attrs==26.1.0
Authlib==1.6.9
bcrypt==5.0.0
billiard==4.2.4
cachetools==7.0.5
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.2
google-cloud-aiplatform>=1.70.0
google-genai==1.73.1
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain==1.2.15
langchain-core==1.2.31
langchain-google-genai==4.2.2
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
lxml==6.0.4
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
neo4j==6.1.0
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
passlib==1.7.4
prompt_toolkit==3.0.52
propcache==0.4.1
py-vapid==1.9.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymilvus==2.6.11
pypdf==6.10.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
pytz==2026.1.post1
pywebpush==2.3.0
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
sentence-transformers==3.0.0
lightrag-hku>=0.1.21
google-generativeai>=0.8.3
networkx>=3.0
zstandard==0.25.0
